Create Web CMS screen, and other changes
Implemented initial version of the Web CMS functionality, using the
android WebKit View. A toolbar is also shown, which allows users to
navigate back and forth, go back to the homepage of CMS, as well as zoom
in and out. A progress bar is also shown while the page is loading,
which animates away once the page is finished loading.

Notes:
- Introduced the `TodayAppFunction` enum class, which replaces the old
  `allAppFunctions` list in `MainActivity.kt`. This was done so that the
  top app bar can distinguish between top level destinations and app
  function pages, and also so navigation using the grid on the home
  screen works a little better.
   - It stores the icon of the app function, as well as it's name
- Created the `TodayScreen` sealed class, used by the top app bar to
  distinguish between top level destinations and app function pages
   - Is either a `TopLevel`, which holds a `TopLevelDestination`, or a
     `AppFunction`, which contains a `TodayAppFunction`
- Improved documentation of various things

Signed-off-by: Erick Howard <needlesslygrim@proton.me>